+++ bdd100k/label/to_mask_test.py
@@ -60,10 +60,11 @@ class TestToMasks(unittest.TestCase):
         bdd100k_config = load_bdd100k_config(task_name)
         convert_func(frames, self.test_out, bdd100k_config.scalabel, 1)
         output_path = os.path.join(self.test_out, output_name)
-        mask = np.asarray(Image.open(output_path))
+        mask = np.asarray(Image.open(output_path), dtype=np.uint8)
 
         gt_mask = np.asarray(
-            Image.open("{}/testcases/{}".format(cur_dir, file_name))
+            Image.open("{}/testcases/{}".format(cur_dir, file_name)),
+            dtype=np.uint8,
         )
 
         self.assertTrue((mask == gt_mask).all())
